feature/#1067 - unifying the tile download code
New classes:
* `CantContinueException`: used to be embedded in `MapTileModuleProviderBase`
* `TileDownloader`: unifies the tile download code, taken from `MapTileDownloader.TileLoader.downloadTile`

Impacted classes:
* `CacheManager`, `MapTileDownloader`: now use a `TileDownloader` in order to download tiles; gently refactored
* `MapTileFilesystemProvider`, `MapTilePreCache`, `MapTileSqlCacheProvider`, `SampleLieFi`: now refer to new class `CantContinueException` instead of `MapTileModuleProviderBase.CantContinueException`
* `MapTileModuleProviderBase`: moved embedded class `CantContinueException` to create an "external/regular" class
