+++ osmdroid-android/src/main/java/org/osmdroid/tileprovider/modules/MapTileFilesystemProvider.java
@@ -142,7 +142,7 @@ public class MapTileFilesystemProvider extends MapTileFileStorageProviderBase {
 				// low memory so empty the queue
 				Log.w(IMapView.LOGTAG, "LowMemoryException downloading MapTile: " + MapTileIndex.toString(pMapTileIndex) + " : " + e);
 				Counters.fileCacheOOM++;
-				throw new MapTileModuleProviderBase.CantContinueException(e);
+				throw new CantContinueException(e);
 			} catch (final Throwable e) {
 				Log.e(IMapView.LOGTAG, "Error loading tile", e);
 				return null;
